move SourceOfTruth out of impl package
