+++ store/src/main/java/com/dropbox/android/external/store4/StoreBuilder.kt
@@ -16,7 +16,6 @@
 package com.dropbox.android.external.store4
 
 import com.dropbox.android.external.store4.impl.RealStore
-import com.dropbox.android.external.store4.impl.SourceOfTruth
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
