+++ store/src/test/java/com/dropbox/android/external/store4/SourceOfTruthWithBarrierTest.kt
@@ -17,7 +17,6 @@ package com.dropbox.android.external.store4
 
 import com.dropbox.android.external.store4.impl.DataWithOrigin
 import com.dropbox.android.external.store4.impl.PersistentSourceOfTruth
-import com.dropbox.android.external.store4.impl.SourceOfTruth
 import com.dropbox.android.external.store4.impl.SourceOfTruthWithBarrier
 import com.dropbox.android.external.store4.testutil.InMemoryPersister
 import com.google.common.truth.Truth.assertThat
